[WORKAROUND] Use WebEngine cookie storage to avoid `::fdatasync`
It appears that using `QSetting` to store cookies leads to frequent
disk cache flushes as `QSattings::setPassword` unconditionally calls
`::fdatasync` on Linux. This may negatively impact performance, and
cause additional wear on SSDs.

This PR is not meant to be the solution but my current workaround,
and I'd like to work out a proper approach.