Add option to control the display of the unread message count on task bar (#649)
In a news reading style with many sources and leaving messages unread
the unread message count becomes irrelevant. We already have an option
to disable the count in the icon tray, but not on the task bar. Added a
checkbox to enable/disable the message count on apps task bar icon /
entry.

Chose to put it into a new tab page as we might add more task bar
functionality (progress update, recent articles, etc..).