ci: only replace dev builds if all jobs succeed
Previously, development binaries were uploaded as soon as their build
job succeeded.

However, because we build for many operating systems (with different
configurations), sometimes a build would succeed on Windows, but not on
Linux, for example.

When that happened, the GitHub Releases page for development builds
would be missing binaries for some OSes, because we were removing _all_
previous binaries before we even invoked CMake.

Now, we only replace the previous development builds if _all_ build
jobs succeed.